release/Sealink-Integrator-Pack/resources/requirements.txt
# requirements.txt - Python packages needed for scripts in this folder
pyserial>=3.5
